make it easy to swap between by election and general election seeds ...
... because I dont want to mess with rails standard rake tasks
